+++ cpp/library/is_prime.cpp
@@ -1,18 +1,18 @@
 // ============how it works=================
 // input:
-  // int N: number to check
+  // ll N: number to check
 // output:
   // bool: is prime or not
 // ========================================
 
 //=============is_prime============================
-bool is_prime(int N)
+bool is_prime(ll N)
 {
   if (N == 1)
   {
     return false;
   }
-  for (int i = 2; i * i <= N; ++i)
+  for (ll i = 2; i * i <= N; ++i)
   {
     if (N % i == 0)
     {
