+++ cpp/library/kruskal.cpp
@@ -3,25 +3,25 @@ using namespace std;
 // ============need input=================
 // MAX_N, MAX_E, V, E
 // output:
-  // int ans: 最小全域木の辺の重みの総和
+  // ll ans: 最小全域木の辺の重みの総和
 // ========================================
 
 
 // ============union_find_tree===============
-int MAX_N = 10000;
-vector<int> par(MAX_N);//MAX_N == num of node
-vector<int> tree_rank(MAX_N);
+ll MAX_N = 10000;
+vector<ll> par(MAX_N);//MAX_N == num of node
+vector<ll> tree_rank(MAX_N);
 
-void init(int n)
+void init(ll n)
 {
-  for (int i = 0; i < n; ++i)
+  for (ll i = 0; i < n; ++i)
   {
     par.at(i) = i;
     tree_rank.at(i) = 0;
   }
 }
 
-int find(int x)
+ll find(ll x)
 {
   if (par.at(x) == x)
   {
@@ -33,7 +33,7 @@ int find(int x)
   }
 }
 
-void unite(int x, int y)
+void unite(ll x, ll y)
 {
   x = find(x);
   y = find(y);
@@ -55,7 +55,7 @@ void unite(int x, int y)
   }
 }
 
-bool same(int x, int y)
+bool same(ll x, ll y)
 {
   return find(x) == find(y);
 }
@@ -63,21 +63,21 @@ bool same(int x, int y)
 
 // ==============kruskal==================
 
-struct edge{int u, v, cost;}; // 辺u, v をつなぐコストが cost
+struct edge{ll u, v, cost;}; // 辺u, v をつなぐコストが cost
 bool comp(const edge& e1, const edge& e2)
 {
   return e1.cost < e2.cost;
 }
-const int MAX_E = 100000;
+const ll MAX_E = 100000;
 edge es[MAX_E];
-int V, E; // V 頂点、E 辺　の数
+ll V, E; // V 頂点、E 辺　の数
 
-int kruskal()
+ll kruskal()
 {
   sort(es, es + E, comp);
   init(V);
-  int ans = 0;
-  for (int i = 0; i < E; ++i)
+  ll ans = 0;
+  for (ll i = 0; i < E; ++i)
   {
     edge e = es[i];
     if (!same(e.u, e.v))
